Add simple extract.acronyms() func, plus docs+tests
