+++ src/textacy/extract/__init__.py
@@ -1,5 +1,5 @@
 from . import keyterms
-from .acronyms import acronyms_and_definitions
+from .acros import acronyms, acronyms_and_definitions
 from .basics import words, ngrams, entities, noun_chunks
 from .kwic import keyword_in_context, KWIC
 from .matches import token_matches, regex_matches
